[FIX] budget_control_purchase: add precommit check budget function
Support for other module can check budget precommit in PO and Actual.
Standard will not check precommit because PO and Actual has next state draft is commit budget.
So, precommit budget is not need.

However, extensnion module like 'base_tier_validation_check_budget' can check budget in state draft.
we will improved base module budget for support precommit check budget

+++ budget_control_purchase/models/account_move_line.py
@@ -35,8 +35,8 @@ class AccountMoveLine(models.Model):
         ):
             inv_state = ml.move_id.state
             move_type = ml.move_id.move_type
-            # Cancel or draft, not commitment line
-            if inv_state != "posted":
+            # State Cancel or draft and not context force commit, not commitment line
+            if not self.env.context.get("force_commit") and inv_state != "posted":
                 self.env["purchase.budget.move"].search(
                     [("move_line_id", "=", ml.id)]
                 ).unlink()
